requirements.txt
pytest==8.3.5
pytest-playwright==0.7.0
allure-pytest==2.13.5
pytest-json-ctrf==0.1.0